requirements.txt
python-telegram-bot==22.1
beautifulsoup4==4.13.4
fastapi==0.115.12
uvicorn==0.34.2